backend-ai/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
pypdf
python-multipart
requests
newspaper3k
beautifulsoup4
lxml_html_clean
groq
mysql-connector-python
passlib[bcrypt]
pydantic[email]
gunicorn
pillow
